% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group4.R
\name{analysis.df}
\alias{analysis.df}
\title{Parse botnet}
\usage{
analysis.df(func_df0)
}
\description{
"main" file which downloads the file if needed
and processes it for a better understanding
}
\examples{
TBD

}
