% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{download.data}
\alias{download.data}
\title{download.data}
\usage{
download.data(site, destfile)
}
\arguments{
\item{site}{URL to get the data from}

\item{destfile}{local path to save downloaded data}
}
\value{
None
}
\description{
Maxmind - Downloads the requested data
}
\examples{
TBD

}
