% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group4.R
\name{generate.df}
\alias{generate.df}
\title{generate.df}
\usage{
generate.df(df, nrows)
}
\arguments{
\item{df}{specific DataFrame to extract info from}

\item{n}{number of rows to extract from the data frame to return}
}
\value{
muestra which will be the requested n first rows of the DF
}
\description{
Returns a DF with first n rows.
}
\examples{
TBD

}
