% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Group4AssigmentPackage.R
\docType{package}
\name{Group4AssigmentPackage}
\alias{Group4AssigmentPackage}
\alias{Group4AssigmentPackage-package}
\title{Group4AssigmentPackage}
\description{
Contains the requested package
}
\author{
Horacio Ignasi
TBD
}
